+++ waggle-dance-core/src/test/java/com/hotels/bdp/waggledance/mapping/service/impl/SimpleFederationStatusServiceTest.java
@@ -17,27 +17,53 @@ package com.hotels.bdp.waggledance.mapping.service.impl;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
+import static org.mockito.Mockito.when;
 
-import org.junit.Rule;
+import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.Mock;
+import org.mockito.runners.MockitoJUnitRunner;
 
+import com.hotels.bdp.waggledance.api.model.FederatedMetaStore;
 import com.hotels.bdp.waggledance.api.model.MetaStoreStatus;
-import com.hotels.beeju.ThriftHiveMetaStoreJUnitRule;
+import com.hotels.bdp.waggledance.client.CloseableThriftHiveMetastoreIface;
+import com.hotels.bdp.waggledance.client.CloseableThriftHiveMetastoreIfaceClientFactory;
 
+@RunWith(MockitoJUnitRunner.class)
 public class SimpleFederationStatusServiceTest {
 
-  public @Rule ThriftHiveMetaStoreJUnitRule metastoreService = new ThriftHiveMetaStoreJUnitRule();
+  private @Mock CloseableThriftHiveMetastoreIfaceClientFactory metaStoreClientFactory;
+  private @Mock CloseableThriftHiveMetastoreIface client;
+  private SimpleFederationStatusService service;
 
-  private SimpleFederationStatusService service = new SimpleFederationStatusService();
+  private final FederatedMetaStore metaStore = FederatedMetaStore.newFederatedInstance("remote", "uri");
 
-  @Test
-  public void metastoreIsAvailable() {
-    assertThat(service.checkStatus(metastoreService.getThriftConnectionUri()), is(MetaStoreStatus.AVAILABLE));
+  @Before
+  public void setUp() {
+    service = new SimpleFederationStatusService(metaStoreClientFactory);
+    when(metaStoreClientFactory.newInstance(metaStore)).thenReturn(client);
   }
 
   @Test
-  public void metastoreIsNotAvailable() {
-    assertThat(service.checkStatus("thrift://unknown:9083"), is(MetaStoreStatus.UNAVAILABLE));
+  public void checkStatusAvailable() throws Exception {
+    when(client.isOpen()).thenReturn(true);
+    MetaStoreStatus status = service.checkStatus(metaStore);
+    assertThat(status, is(MetaStoreStatus.AVAILABLE));
+  }
+
+  @Test
+  public void checkStatusUnavailable() throws Exception {
+    when(client.isOpen()).thenReturn(false);
+    MetaStoreStatus status = service.checkStatus(metaStore);
+    assertThat(status, is(MetaStoreStatus.UNAVAILABLE));
+  }
+
+  @Test
+  public void checkStatusUnavailableViaException() throws Exception {
+    when(client.isOpen()).thenThrow(new RuntimeException("oh no metastore down!"));
+    MetaStoreStatus status = service.checkStatus(metaStore);
+    assertThat(status, is(MetaStoreStatus.UNAVAILABLE));
   }
 
 }